+++ ixmp/testing.py
@@ -120,12 +120,12 @@ def create_local_testdb(db_path, data_path, db='ixmptest'):
     are copied from *data_path*.
     """
     # Copy test database
-    dst = db_path / 'testdb'
+    dst = Path(db_path) / 'testdb'
     # str() here is for py2 compatibility
     shutil.copytree(str(data_path), str(dst))
 
     # Create properties file
-    props = (data_path / 'test.properties_template').read_text()
+    props = (Path(data_path) / 'test.properties_template').read_text()
     test_props = dst / 'test.properties'
     test_props.write_text(props.format(here=str(dst).replace("\\", "/"),
                           db=db))
@@ -205,7 +205,7 @@ def make_dantzig(mp, solve=False):
 
     if solve:
         # Solve the model using the GAMS code provided in the `tests` folder
-        scen.solve(model=str(solve / 'transport_ixmp'),
+        scen.solve(model=str(Path(solve) / 'transport_ixmp'),
                    case='transport_standard')
 
     # add timeseries data for testing `clone(keep_solution=False)`