+++ rixmp/tests/testthat.R
@@ -1,0 +1,5 @@
+library(testthat)
+library(rixmp)
+
+test_check("rixmp")
+ixmp <- import('ixmp')